+++ .storybook/themes.ts
@@ -107,9 +107,19 @@ export const themes: Theme[] = [
   {
     id: 'neon',
     name: 'NEON',
-    fonts: ['https://use.typekit.net/ghy4rhf.css'],
+    fonts: ['https://use.typekit.net/ghy4rhf.css','https://use.typekit.net/qod6igq.css'],
     setup: async () => {
       const { setup } = await import('@sl-design-system/neon');
+
+      setup();
+    }
+  },
+  {
+    id: 'nowa-era',
+    name: 'Nowa Era',
+    fonts: ['https://use.typekit.net/ghy4rhf.css','https://use.typekit.net/qod6igq.css'],
+    setup: async () => {
+      const { setup } = await import('@sl-design-system/nowa-era');
 
       setup();
     }