+++ packages/components/select/src/select-option.ts
@@ -1,4 +1,5 @@
 import type { CSSResultGroup, TemplateResult } from 'lit';
+import type { SelectSize } from './select.js';
 import type { FormControlValue } from '@sl-design-system/shared';
 import { observe } from '@sl-design-system/shared';
 import { LitElement, html } from 'lit';
@@ -9,31 +10,22 @@ export class SelectOption extends LitElement {
   /** @private */
   static override styles: CSSResultGroup = styles;
 
-  /** Whether the option item is selected*/
-  @property() value?: FormControlValue;
+  /** @ignore The size of the select, is set by the select component. */
+  @property({ reflect: true }) size: SelectSize = 'md';
 
-  /** Whether the option item is selected*/
-  @property({ reflect: true, type: Boolean }) selected = false;
-
-  /** Whether the option item is disabled*/
-  @property({ reflect: true, type: Boolean }) disabled = false;
-
-  /** Whether the content of the option item is a node*/
+  /** @ignore Whether the content of the option item is a node*/
   @property({ reflect: true }) contentType?: 'string' | 'element';
 
-  @property({ reflect: true }) size: { width: number; height: number } = { width: 0, height: 0 };
+  /** The value for the option item, to be used in forms.*/
+  @property() value?: FormControlValue;
 
-  #observer?: ResizeObserver;
+  /** Whether the option item is selected. */
+  @property({ reflect: true, type: Boolean }) selected = false;
 
-  /** Get the selected tab button, or the first tab button. */
-  get #tabIndex(): string | null {
-    return this.getAttribute('tabIndex');
-  }
+  /** Whether the option item is disabled. */
+  @property({ reflect: true, type: Boolean }) disabled = false;
 
-  /**
-   * Apply accessible attributes and values to the tab button.
-   * Observe the selected property if it changes
-   */
+  /** @ignore Apply accessible attributes and values to the option. Observe the selected property if it changes */
   @observe('selected')
   protected handleSelectionChange(): void {
     this.setAttribute('aria-selected', this.selected ? 'true' : 'false');
@@ -45,32 +37,11 @@ export class SelectOption extends LitElement {
     this.setAttribute('role', 'option');
   }
 
-  override disconnectedCallback(): void {
-    super.disconnectedCallback();
-
-    this.#observer?.disconnect();
-  }
-
-  override firstUpdated(): void {
-    this.#observer = new ResizeObserver(m => this.#handleResize(m));
-    this.#observer?.observe(this);
-  }
-
   override render(): TemplateResult {
     return html`<slot @slotchange=${this.#onSlotchange}></slot>`;
   }
 
   async #onSlotchange(event: Event & { target: HTMLSlotElement }): Promise<void> {
     this.contentType = event.target.assignedNodes()[0].nodeType === 1 ? 'element' : 'string';
-    this.size = { width: this.getBoundingClientRect().width, height: this.getBoundingClientRect().height };
-  }
-
-  #handleResize(mutations: ResizeObserverEntry[]): void {
-    mutations.forEach(mutation => {
-      this.size = {
-        width: mutation.borderBoxSize[0].inlineSize,
-        height: mutation.borderBoxSize[0].blockSize
-      };
-    });
   }
 }