+++ website/site.eleventy.cjs
@@ -95,7 +95,7 @@ module.exports = function(eleventyConfig) {
 
   let mdIt = markdownIt({
     html: true,
-    breaks: true,
+    breaks: false,
     linkify: true,
   })
     .disable('code')
@@ -121,7 +121,7 @@ module.exports = function(eleventyConfig) {
     .addPassthroughCopy('./src/site/assets')
     .addPassthroughCopy({ './../packages/themes/sanoma-learning': `styles/slds-sanoma-learning` })
     .addPassthroughCopy({ './../packages/tokens/src/sanoma-learning/*.json': `_data` });
-   
+
   const NOT_FOUND_PATH = `${outputFolder}/site/404.html`;
 
   eleventyConfig.setBrowserSyncConfig({